Merged in Cyril's AutoMasking_SNR branch. Added peak-weighting image option to HMP.

+++ DDFacet/DDF.py
@@ -45,15 +45,17 @@ from DDFacet.Other import MyPickle
 from DDFacet.Parset import MyOptParse
 from DDFacet.Other import MyLogger
 from DDFacet.Other import ModColor
+from DDFacet.Other import Exceptions
 from DDFacet.ToolsDir import ModFFTW
 from DDFacet.Other import ClassTimeIt
 from DDFacet.Other import Multiprocessing
 import SkyModel.Other.ModColor   # because it's duplicated there
 from DDFacet.Other import progressbar
-from DDFacet.Other.AsyncProcessPool import APP
+from DDFacet.Other.AsyncProcessPool import APP, WorkerProcessError
 import DDFacet.cbuild.Gridder._pyArrays as _pyArrays
 from DDFacet.version import __version__
 log = None
+# from version import __version__
 
 import numpy as np
 
@@ -84,8 +86,9 @@ def report_version():
 # # Catch numpy warning
 # np.seterr(all='raise')
 # import warnings
-# with warnings.catch_warnings():
-#     warnings.filterwarnings('error')
+# warnings.filterwarnings('error')
+# #with warnings.catch_warnings():
+# #    warnings.filterwarnings('error')
 # # ##############################
 
 '''
@@ -141,12 +144,15 @@ def test():
 def main(OP=None, messages=[]):
     if OP is None:
         OP = MyPickle.Load(SaveFile)
+        print "Using settings from %s, then command line."%SaveFile
 
     DicoConfig = OP.DicoConfig
 
     ImageName = DicoConfig["Output"]["Name"]
     if not ImageName:
-        raise ValueError("--Output-Name not specified")
+        raise Exceptions.UserInputError("--Output-Name not specified, can't continue.")
+    if not DicoConfig["Data"]["MS"]:
+        raise Exceptions.UserInputError("--Data-MS not specified, can't continue.")
 
     # create directory if it exists
     dirname = os.path.dirname(ImageName)
@@ -169,6 +175,15 @@ def main(OP=None, messages=[]):
             logo.print_logo()
         for msg in messages:
             print>> log, msg
+
+    if DicoConfig["Debug"]["Pdb"] == "always":
+        print>>log, "--Debug-Pdb=always: unexpected errors will be dropped into pdb"
+        Exceptions.enable_pdb_on_error(ModColor.Str("DDFacet has encountered an unexpected error. Dropping you into pdb for a post-mortem.\n" +
+                                           "(This is because you're running with --Debug-Pdb set to 'always'.)"))
+    elif DicoConfig["Debug"]["Pdb"] == "auto" and not DicoConfig["Log"]["Boring"]:
+        print>>log, "--Debug-Pdb=auto and not --Log-Boring: unexpected errors will be dropped into pdb"
+        Exceptions.enable_pdb_on_error(ModColor.Str("DDFacet has encountered an unexpected error. Dropping you into pdb for a post-mortem.\n" +
+            "(This is because you're running with --Debug-Pdb set to 'auto' and --Log-Boring is off.)"))
 
     # print current options
     OP.Print(dest=log)
@@ -207,15 +222,15 @@ def main(OP=None, messages=[]):
     # write parset
     OP.ToParset("%s.parset"%ImageName)
 
-    Mode = DicoConfig["Image"]["Mode"]
+    Mode = DicoConfig["Output"]["Mode"] 
 
     # data machine initialized for all cases except PSF-only mode
     # psf machine initialized for all cases except Predict-only mode
     Imager = ClassDeconvMachine.ClassImagerDeconv(GD=DicoConfig, 
                                                   BaseName=ImageName,
-                                                  predict_only=(Mode == "Predict"),
+                                                  predict_only=(Mode == "Predict" or Mode == "Subtract"),
                                                   data=(Mode != "PSF"), 
-                                                  psf=(Mode != "Predict" and Mode != "Dirty"),
+                                                  psf=(Mode != "Predict" and Mode != "Dirty" and Mode != "Subtract"),
                                                   readcol=(Mode != "Predict" and Mode != "PSF"),
                                                   deconvolve=("Clean" in Mode))
 
@@ -237,6 +252,8 @@ def main(OP=None, messages=[]):
         if sparsify and isinstance(sparsify, list):
             sparsify = sparsify[0]
         Imager.MakePSF(sparsify=sparsify)
+    elif "RestoreAndShift" == Mode:
+        Imager.RestoreAndShift()
 
     # # open default viewer, these options should match those in
     # # ClassDeconvMachine if changed:
@@ -317,7 +334,6 @@ def main(OP=None, messages=[]):
     #         print>>log, ModColor.Str(
     #             "\nDon't understand %s, not opening that image\n" %
     #             img, col="yellow")
-    Multiprocessing.cleanupShm()
 
 if __name__ == "__main__":
     #os.system('clear')
@@ -374,7 +390,7 @@ if __name__ == "__main__":
         OP.ExitWithError("Incorrect number of arguments. Use -h for help.")
         sys.exit(1)
 
-    retcode = 0
+    retcode = report_error = 0
     try:
         main(OP, messages)
         print>>log, ModColor.Str(
@@ -385,25 +401,40 @@ if __name__ == "__main__":
         print>>log, ModColor.Str("DDFacet interrupted by Ctrl+C", col="red")
         APP.terminate()
         retcode = 1 #Should at least give the command line an indication of failure
+    except Exceptions.UserInputError:
+        print>> log, ModColor.Str(sys.exc_info()[1], col="red")
+        print>> log, ModColor.Str("There was a problem with some user input. See messages above for an indication.")
+        APP.terminate()
+        retcode = 1  # Should at least give the command line an indication of failure
+    except WorkerProcessError:
+        print>> log, ModColor.Str("A worker process has died on us unexpectedly. This probably indicates a bug:")
+        print>> log, ModColor.Str("  the original underlying error may be reported in the log [possibly far] above.")
+        report_error = True
     except:
         print>>log, traceback.format_exc()
+        if sys.exc_info()[0] is not WorkerProcessError and Exceptions.is_pdb_enabled():
+            raise
+        report_error = True
 
+    if report_error:
         logfileName = MyLogger.getLogFilename()
         logfileName = logfileName if logfileName is not None else "[file logging is not enabled]"
+        print>> log, ""
         print>> log, ModColor.Str(
-            "There was a problem after %s, if you think this is a bug please open an "
-            "issue, quote your version of DDFacet and attach your logfile" %
-            T.timehms(), col="red")
+            "There was a problem after %s; if you think this is a bug please open an issue, "%
+            T.timehms(), col = "red")
+        print>> log, ModColor.Str("  quote your version of DDFacet and attach your logfile.", col="red")
         print>> log, ModColor.Str(
             "You are using DDFacet revision: %s" %
             version, col="red")
         print>> log, ModColor.Str(
             "Your logfile is available here: %s" %
             logfileName, col="red")
-        print>>log, traceback_msg
+        # print>>log, traceback_msg
         # Should at least give the command line an indication of failure
         APP.terminate()
         retcode = 1 # Should at least give the command line an indication of failure
 
     APP.shutdown()
+    Multiprocessing.cleanupShm()
     sys.exit(retcode)
